+++ src/biz/bokhorst/xprivacy/XPrivacy.java
@@ -13,7 +13,6 @@ import android.content.Context;
 import android.os.Binder;
 import android.os.Build;
 import android.os.Process;
-import android.os.SystemClock;
 import android.util.Log;
 
 import de.robv.android.xposed.IXposedHookLoadPackage;