+++ CHANGELOG.md
@@ -16,13 +16,15 @@ Test and beta releases will have experimental functions enabled by default.
 
 **Next release**
 
+[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
+
+**Version 1.11.11 TEST**
+
 * Fixed restricting locations for some Android versions ([issue](https://github.com/M66B/XPrivacy/issues/1102))
 * Fixed restricting incoming phone number for some Android versions
 * Fixed removing location and phone state listener
 * Fixed location client always restricting locations
 * Fixed fake value for SIM (ICC) operator name
-
-[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 1.11.10 TEST**